add repo mysql ping implementation
